features: wifi: make CFG/MAC80211 modules
We use CFG80211 with the new regulatory.db mechanism. This loads
that file using the firmware loading mechanism. If the module is
built-in, the kernel complains since it can't load regulatory.db:
  kernel: platform regulatory.0: Direct firmware load for regulatory.db failed with error -2

Configuring CFG80211 as a module allows the kernel to load
regulatory.db using the firmware loading mechanism. That also
helps keeping the kernel smaller for users which do not use WiFi.

Also drop the legacy CRDA support since we anyway don't deploy the
necessary user-space tools with TorizonCore.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>