features: media: disable autoselect
Disabling autoselect is required to be able to select sensors such
as OV5640 directly. This will make enabling standard multimedia
devices such as USB DVB-T receivers a bit a hassle: One will need
to make sure to include all subdevices as it would be selected by
MEDIA_SUBDRV_AUTOSELECT (see drivers/media/usb/dvb-usb-v2/Kconfig).
But it is the only way to select sensors such as OV5640 directly
as far as I know.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>