imx8: enable CONFIG_FEC_UIO as module
Enable the 10/100 Fast ethernet controller on Freescale i.MX 8M Mini
processor as a module.

This is configured as built-in in the kernel, but we have to change it
to a module since it depends on CONFIG_UIO that is already configured as
a module in a common configuration file.

Related-to: TOR-2424

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>