sound: change a few configs to build as modules
Those config options cannot be configured as built-in because they
depend on other options that are already configured as modules. This
prevent warnings from OE when building the kernel.

And indeed, doesn't make sense to have them as built in, since not all
customers will use audio. In theory, every driver that is not required
to boot the device should not be built-in to not bloat the kernel.

Related-to: TOR-2424

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>