base: enable SWAP
This is not enabling SWAP in TorizonCore, just making it available in
the kernel.

And since SWAP is a kernel feature not related to the BSP, we are safe
to enable it for all supported hardware in the kernel cache.

This is also required to prevent the following warning message from
Docker:

"WARNING: No swap limit support"

Related-to: TOR-2479

Signed-off-by: Sergio Prado <sergio.prado@toradex.com>