display: enable necessary display modules as builtin
Due to Torizon using initramfs, where plymouth starts up and shows the
booting splash screen, we need a handful of modules available at that
time, be it as modules to be inserted or as builtin.

On BSP 7, they changed the way they keep their Kernel configs, and with
that some configs changed from builtin to module, impacting the splash
screen on most of our devices.

So, instead of adding manually those modules to the initramfs, we'll be
changing them to builtin on our builds, since there's no downside in
doing so.

Related-to: TOR-3594

Signed-off-by: Eduardo Ferreira <eduardo.barbosa@toradex.com>